requirements.txt
anthropic==0.47.0
openai==1.59.7
google-generativeai==0.8.3
yt-dlp==2024.12.13
python-dotenv==1.0.0
pydantic==2.10.5
